List and watch by namespace
