+++ internal/controllers/replication/symphony.go
@@ -41,7 +41,7 @@ func (c *symphonyController) Reconcile(ctx context.Context, req ctrl.Request) (c
 	ctx = logr.NewContext(ctx, logger)
 
 	existing := &apiv1.CompositionList{}
-	err = c.client.List(ctx, existing, client.MatchingFields{
+	err = c.client.List(ctx, existing, client.InNamespace(symph.Namespace), client.MatchingFields{
 		manager.IdxCompositionsBySymphony: symph.Name,
 	})
 	if err != nil {
