+++ internal/controllers/synthesis/lifecycle.go
@@ -70,7 +70,7 @@ func (c *podLifecycleController) Reconcile(ctx context.Context, req ctrl.Request
 	// Delete any unnecessary pods
 	pods := &corev1.PodList{}
 	err = c.client.List(ctx, pods, client.InNamespace(c.config.PodNamespace), client.MatchingFields{
-		manager.IdxPodsByComposition: comp.Name,
+		manager.IdxPodsByComposition: manager.PodByCompIdxValueFromComp(comp),
 	})
 	if err != nil {
 		return ctrl.Result{}, fmt.Errorf("listing pods: %w", err)
